% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_table.R
\name{calculate_table}
\alias{calculate_table}
\title{calculate percentage tables for comment themes}
\usage{
calculate_table(
  table_data,
  code_column,
  category_table,
  join_lookup,
  count_column,
  click_column = NULL
)
}
\arguments{
\item{table_data}{A dataframe}

\item{code_column}{String with name of column to summarise.}

\item{category_table}{The table to join with.}

\item{join_lookup}{A named vector giving the names to join on}

\item{count_column}{String with the name of the column to count in the joined table}

\item{click_column}{A string with the name of the category that has been clicked, if any}
}
\value{
a dataframe with category name, n, and %
}
\description{
calculate percentage tables for comment themes
}
